requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
uvicorn==0.22.0  # Add uvicorn with an appropriate version
